Start migrating some query endpoints to new pattern
